Break out ModelResolver object
