+++ modules/scala-support/src/main/scala/dev/guardrail/generators/scala/circe/CirceRefinedProtocolGenerator.scala
@@ -14,6 +14,7 @@ import scala.reflect.runtime.universe.typeTag
 import dev.guardrail.core
 import dev.guardrail.core.extract.Default
 import dev.guardrail.core.extract.{ DataRedaction, EmptyValueIsNull }
+import dev.guardrail.core.resolvers.ModelResolver
 import dev.guardrail.core.{ DataRedacted, DataVisible, EmptyIsEmpty, EmptyIsNull, LiteralRawType, Mappish, ReifiedRawType, SupportDefinition, Tracker }
 import dev.guardrail.generators.ProtocolGenerator.{ WrapEnumSchema, wrapNumberEnumSchema, wrapObjectEnumSchema, wrapStringEnumSchema }
 import dev.guardrail.generators.protocol.{ ClassChild, ClassHierarchy, ClassParent }
@@ -218,7 +219,7 @@ class CirceRefinedProtocolGenerator private (circeVersion: CirceModelGenerator, 
                 components
               )
               customTypeName <- SwaggerUtil.customTypeName(x)
-              (declType, _)  <- SwaggerUtil.determineTypeName[ScalaLanguage, Target](x, Tracker.cloneHistory(x, customTypeName), components)
+              (declType, _)  <- ModelResolver.determineTypeName[ScalaLanguage, Target](x, Tracker.cloneHistory(x, customTypeName), components)
               alias          <- typeAlias(formattedClsName, declType)
             } yield enum.orElse(model).getOrElse(alias)
           )
@@ -238,7 +239,7 @@ class CirceRefinedProtocolGenerator private (circeVersion: CirceModelGenerator, 
                 components
               )
               customTypeName <- SwaggerUtil.customTypeName(x)
-              (declType, _)  <- SwaggerUtil.determineTypeName[ScalaLanguage, Target](x, Tracker.cloneHistory(x, customTypeName), components)
+              (declType, _)  <- ModelResolver.determineTypeName[ScalaLanguage, Target](x, Tracker.cloneHistory(x, customTypeName), components)
               alias          <- typeAlias(formattedClsName, declType)
             } yield enum.orElse(model).getOrElse(alias)
           )
@@ -246,7 +247,7 @@ class CirceRefinedProtocolGenerator private (circeVersion: CirceModelGenerator, 
             for {
               formattedClsName <- formatTypeName(clsName)
               customTypeName   <- SwaggerUtil.customTypeName(x)
-              (declType, _)    <- SwaggerUtil.determineTypeName[ScalaLanguage, Target](x, Tracker.cloneHistory(x, customTypeName), components)
+              (declType, _)    <- ModelResolver.determineTypeName[ScalaLanguage, Target](x, Tracker.cloneHistory(x, customTypeName), components)
               res              <- typeAlias(formattedClsName, declType)
             } yield res
           )
@@ -350,7 +351,7 @@ class CirceRefinedProtocolGenerator private (circeVersion: CirceModelGenerator, 
     for {
       enum          <- extractEnum(schema.map(wrapEnumSchema))
       customTpeName <- SwaggerUtil.customTypeName(schema)
-      (tpe, _)      <- SwaggerUtil.determineTypeName(schema, Tracker.cloneHistory(schema, customTpeName), components)
+      (tpe, _)      <- ModelResolver.determineTypeName(schema, Tracker.cloneHistory(schema, customTpeName), components)
       fullType      <- selectType(NonEmptyList.ofInitLast(dtoPackage, clsName))
       res           <- enum.traverse(validProg(_, tpe, fullType))
     } yield res
@@ -413,7 +414,7 @@ class CirceRefinedProtocolGenerator private (circeVersion: CirceModelGenerator, 
           typeName <- formatTypeName(name).map(formattedName => NonEmptyList.of(hierarchy.name, formattedName))
           propertyRequirement = getPropertyRequirement(prop, requiredFields.contains(name), defaultPropertyRequirement)
           customType <- SwaggerUtil.customTypeName(prop)
-          resolvedType <- SwaggerUtil
+          resolvedType <- ModelResolver
             .propMeta[ScalaLanguage, Target](
               prop,
               components
@@ -657,7 +658,7 @@ class CirceRefinedProtocolGenerator private (circeVersion: CirceModelGenerator, 
             typeName              <- formatTypeName(name).map(formattedName => getClsName(name).append(formattedName))
             tpe                   <- selectType(typeName)
             maybeNestedDefinition <- processProperty(name, schema)
-            resolvedType          <- SwaggerUtil.propMetaWithName(tpe, schema, components)
+            resolvedType          <- ModelResolver.propMetaWithName(tpe, schema, components)
             customType            <- SwaggerUtil.customTypeName(schema)
             propertyRequirement = getPropertyRequirement(schema, requiredFields.contains(name), defaultPropertyRequirement)
             defValue  <- defaultValue(typeName, schema, propertyRequirement, definitions)
@@ -772,7 +773,7 @@ class CirceRefinedProtocolGenerator private (circeVersion: CirceModelGenerator, 
       tpe <- model.fold[Target[scala.meta.Type]](objectType(None)) { m =>
         for {
           tpeName       <- SwaggerUtil.customTypeName[ScalaLanguage, Target, Tracker[ObjectSchema]](m)
-          (declType, _) <- SwaggerUtil.determineTypeName[ScalaLanguage, Target](m, Tracker.cloneHistory(m, tpeName), components)
+          (declType, _) <- ModelResolver.determineTypeName[ScalaLanguage, Target](m, Tracker.cloneHistory(m, tpeName), components)
         } yield declType
       }
       res <- typeAlias(clsName, tpe)
@@ -801,7 +802,7 @@ class CirceRefinedProtocolGenerator private (circeVersion: CirceModelGenerator, 
       Sw: SwaggerTerms[ScalaLanguage, Target]
   ): Target[ProtocolElems[ScalaLanguage]] =
     for {
-      deferredTpe <- SwaggerUtil.modelMetaType(arr, components)
+      deferredTpe <- ModelResolver.modelMetaType(arr, components)
       tpe         <- extractArrayType(deferredTpe, concreteTypes)
       ret         <- typeAlias(clsName, tpe)
     } yield ret
